changed the superclass
Closes #47811
