+++ tests/base/src/test/java/org/keycloak/tests/sessionlimits/KcOidcUserSessionLimitsBrokerTest.java
@@ -16,148 +16,62 @@
  */
 package org.keycloak.tests.sessionlimits;
 
-import org.keycloak.authentication.authenticators.sessionlimits.UserSessionLimitsAuthenticatorFactory;
 import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
 import org.keycloak.broker.oidc.OIDCIdentityProviderFactory;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.testframework.annotations.InjectRealm;
 import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.injection.LifeCycle;
-import org.keycloak.testframework.oauth.OAuthClient;
-import org.keycloak.testframework.oauth.annotations.InjectOAuthClient;
 import org.keycloak.testframework.realm.ClientBuilder;
 import org.keycloak.testframework.realm.IdentityProviderBuilder;
 import org.keycloak.testframework.realm.ManagedRealm;
 import org.keycloak.testframework.realm.RealmBuilder;
 import org.keycloak.testframework.realm.RealmConfig;
 import org.keycloak.testframework.realm.UserBuilder;
-import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
-import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
-import org.keycloak.testframework.ui.annotations.InjectPage;
-import org.keycloak.testframework.ui.annotations.InjectWebDriver;
-import org.keycloak.testframework.ui.page.ErrorPage;
-import org.keycloak.testframework.ui.page.LoginPage;
-import org.keycloak.testframework.ui.webdriver.ManagedWebDriver;
 
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 import static org.keycloak.broker.oidc.OAuth2IdentityProviderConfig.TOKEN_ENDPOINT_URL;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.CONSUMER_CLIENT_ID;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.CONSUMER_CLIENT_SECRET;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.CONSUMER_REALM;
-import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.ERROR_TO_DISPLAY;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.PROVIDER_REALM;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.USER_EMAIL;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.USER_LOGIN;
 import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.USER_PASSWORD;
-import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.assertSessionCount;
-import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.cleanupBeforeTest;
-import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.configurePostBrokerFlow;
-import static org.keycloak.tests.sessionlimits.UserSessionLimitsUtil.deleteAllCookies;
 
 @KeycloakIntegrationTest
-public class KcOidcUserSessionLimitsBrokerTest {
+public class KcOidcUserSessionLimitsBrokerTest extends AbstractUserSessionLimitsBrokerTest {
 
     private static final String IDP_ALIAS = "kc-oidc-idp";
     private static final String BROKER_CLIENT_ID = "brokerapp";
     private static final String BROKER_CLIENT_SECRET = "secret";
 
-    @InjectRealm(ref = "provider", config = ProviderRealmConfig.class)
+    @InjectRealm(ref = "provider", lifecycle = LifeCycle.METHOD, config = ProviderRealmConfig.class)
     ManagedRealm providerRealm;
 
-    @InjectRealm(ref = "consumer", config = ConsumerRealmConfig.class)
+    @InjectRealm(ref = "consumer", lifecycle = LifeCycle.METHOD, config = ConsumerRealmConfig.class)
     ManagedRealm consumerRealm;
 
-    @InjectRunOnServer(realmRef = "consumer")
-    RunOnServerClient runOnServer;
-
-    @InjectOAuthClient(lifecycle = LifeCycle.METHOD)
-    OAuthClient oauth;
-
-    @InjectWebDriver
-    ManagedWebDriver driver;
-
-    @InjectPage
-    LoginPage loginPage;
-
-    @InjectPage
-    ErrorPage errorPage;
-
-    @BeforeEach
-    public void setup() {
-        cleanupBeforeTest(driver, consumerRealm, providerRealm, runOnServer);
+    @Override
+    protected String getIdpAlias() {
+        return IDP_ALIAS;
     }
 
-    @Test
-    public void testSessionCountExceededAndNewSessionDeniedFirstBrokerLoginFlow() {
-        runOnServer.run(configurePostBrokerFlow(CONSUMER_REALM, IDP_ALIAS,
-                UserSessionLimitsAuthenticatorFactory.DENY_NEW_SESSION, "0", "1"));
-
-        logInAsUserInIDPForFirstTime();
-        deleteAllCookies(driver, consumerRealm);
-        deleteAllCookies(driver, providerRealm);
-        logInAsUserInIDP();
-
-        errorPage.assertCurrent();
-        Assertions.assertEquals(ERROR_TO_DISPLAY, errorPage.getError());
+    @Override
+    protected ManagedRealm getConsumerRealm() {
+        return consumerRealm;
     }
 
-    @Test
-    public void testSessionCountExceededAndOldestSessionRemovedFirstBrokerLoginFlow() {
-        runOnServer.run(configurePostBrokerFlow(CONSUMER_REALM, IDP_ALIAS,
-                UserSessionLimitsAuthenticatorFactory.TERMINATE_OLDEST_SESSION, "0", "1"));
-
-        logInAsUserInIDPForFirstTime();
-        deleteAllCookies(driver, consumerRealm);
-        deleteAllCookies(driver, providerRealm);
-        logInAsUserInIDP();
-
-        Assertions.assertTrue(oauth.parseLoginResponse().isSuccess());
-        runOnServer.run(assertSessionCount(CONSUMER_REALM, USER_LOGIN, 1));
+    @Override
+    protected ManagedRealm getProviderRealm() {
+        return providerRealm;
     }
 
-    @Test
-    public void testRealmSessionCountExceededAndNewSessionDeniedFirstBrokerLoginFlow() {
-        runOnServer.run(configurePostBrokerFlow(CONSUMER_REALM, IDP_ALIAS,
-                UserSessionLimitsAuthenticatorFactory.DENY_NEW_SESSION, "1", "0"));
-
-        logInAsUserInIDPForFirstTime();
-        deleteAllCookies(driver, consumerRealm);
-        deleteAllCookies(driver, providerRealm);
-        logInAsUserInIDP();
-
-        errorPage.assertCurrent();
-        Assertions.assertEquals(ERROR_TO_DISPLAY, errorPage.getError());
-    }
-
-    @Test
-    public void testRealmSessionCountExceededAndOldestSessionRemovedFirstBrokerLoginFlow() {
-        runOnServer.run(configurePostBrokerFlow(CONSUMER_REALM, IDP_ALIAS,
-                UserSessionLimitsAuthenticatorFactory.TERMINATE_OLDEST_SESSION, "1", "0"));
-
-        logInAsUserInIDPForFirstTime();
-        deleteAllCookies(driver, consumerRealm);
-        deleteAllCookies(driver, providerRealm);
-        logInAsUserInIDP();
-
-        Assertions.assertTrue(oauth.parseLoginResponse().isSuccess());
-        runOnServer.run(assertSessionCount(CONSUMER_REALM, USER_LOGIN, 1));
-    }
-
-    private void logInAsUserInIDPForFirstTime() {
+    @Override
+    protected void logInAsUserInIDPForFirstTime() {
         logInAsUserInIDP();
         Assertions.assertTrue(oauth.parseLoginResponse().isSuccess());
-    }
-
-    private void logInAsUserInIDP() {
-        oauth.realm(CONSUMER_REALM).client(CONSUMER_CLIENT_ID, CONSUMER_CLIENT_SECRET).openLoginForm();
-        loginPage.assertCurrent();
-        loginPage.clickSocial(IDP_ALIAS);
-        loginPage.assertCurrent();
-        loginPage.fillLogin(USER_LOGIN, USER_PASSWORD);
-        loginPage.submit();
     }
 
     public static class ProviderRealmConfig implements RealmConfig {
